[Feat] 댓글 입력창 컴포넌트로 빼기 Close #215
